+++ src/variant_converters/packed_int32_array_variant_converter.cpp
@@ -17,7 +17,7 @@ void YAMLPackedInt32ArrayVariantConverter::encode(ryml::NodeRef &node, const Var
 	// Get template style if it exists (key "_template" is a convention for shared array item styling)
 	YAMLStyle::View template_style = style.get_template_style();
 
-	for (int i = 0; i < array.size(); ++i) {
+	for (int64_t i = 0; i < array.size(); ++i) {
 		ryml::NodeRef value_node = node.append_child();
 
 		YAMLStyle::View item_style;
